==> 02-data_wrangling.ipynb <==
# Data Wrangling & Feature Engineering

import pandas as pd
import numpy as np

user_df = pd.read_csv('User_Listening_History.csv')
music_df = pd.read_csv('Music_Info.csv')

music_df.head()

user_df.head()

### 1. Wrangling & Cleaning

# the training RMSE lowered with lower playcount limit
# but RMSE is not the best metric for knn

# change pc_limit to change the amount of playcounts
pc_limit = 5

# keep interactions with less than pc_limit playcounts
user_df = user_df[user_df['playcount'] <= pc_limit]

#### Genre encoding

# check for nulls
music_df.isnull().sum()

print(music_df['genre'].unique())

music_df['tags'].value_counts()

# genre to lowercase
music_df['genre'] = music_df['genre'].str.lower()

# tags and genre null values empty string
music_df['tags'].fillna('', inplace=True)
music_df['genre'].fillna('', inplace=True)

def concatenate_genre_tags(row):
    genre = row['genre'].strip()
    tags = row['tags'].strip()

    if genre and genre not in tags:
        combined = f"{genre}, {tags}" if tags else genre
    else:
        combined = tags
    return combined


# make column tagsplus that is tags plus genre with comma and space
music_df['tagsplus'] = music_df.apply(concatenate_genre_tags, axis=1)

music_df['tagsplus']

# drop rows where tagsplus is empty string
music_df = music_df[music_df['tagsplus'] != '']

# for tagsplus count the amount of empty strings
music_df['tagsplus'].isnull().sum()

len(music_df)

### 2. Feature Engineering

from sklearn.preprocessing import StandardScaler, MultiLabelBinarizer
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline


# drop genre, tags and spotify fields
music_df.drop(columns=['genre', 'tags', 'spotify_preview_url', 'spotify_id'], inplace=True)

#### 2.1 Numerical scaling

# numeric features and transformer
numeric_features = ['year', 'duration_ms', 'danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness', 
                    'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo', 'time_signature']
numeric_transformer = Pipeline(steps=[
    ('scaler', StandardScaler())])


preprocessor = ColumnTransformer(
    transformers=[
        ('num', numeric_transformer, numeric_features)])

numeric_processed = preprocessor.fit_transform(music_df)

numeric_processed.shape

#### 2.2 Genre one-hot encoding

# tagsplus string to list
tagsplus_df = music_df['tagsplus'].apply(lambda x: x.split(', ') if isinstance(x, str) else [])

# use multilabelbinarizer to encode tagsplus
mlb = MultiLabelBinarizer()

genre_encoded = mlb.fit_transform(tagsplus_df)

genre_df = pd.DataFrame(genre_encoded, columns=mlb.classes_, index=tagsplus_df.index)

genre_df.head()

# re-merge list of numeric_features with genre_df dataframe
features_df = pd.concat([pd.DataFrame(numeric_processed, columns=numeric_features, index=music_df.index), genre_df], axis=1)


# add track_id, name and artist to features_df
features_df = pd.concat([music_df[['track_id', 'name', 'artist']], features_df], axis=1)


len(features_df)

features_df.head()

#### 2.3 User playcounts scaling

from sklearn.preprocessing import MinMaxScaler

# Initialize the scaler, specifying the range. scikit-surprise default is (1, 5)
scaler = MinMaxScaler(feature_range=(1, 5))

# Transform the playcount column
user_df['playcount_scale'] = scaler.fit_transform(user_df['playcount'].values.reshape(-1, 1))

# replace playcount w scaled
user_df['playcount'] = user_df['playcount_scale']
user_df.drop(columns=['playcount_scale'], inplace=True)

user_df.head()

#### 2.4 Dataset consistency

# remove tracks from user_df that are not in music_df
user_df = user_df[user_df['track_id'].isin(features_df['track_id'])]

# remove tracks from music_df that are not in user_df
features_df = features_df[features_df['track_id'].isin(user_df['track_id'])]

# sanity check
len(features_df['track_id'].unique()) == len(user_df['track_id'].unique())

# reset indices
user_df.reset_index(drop=True, inplace=True)
features_df.reset_index(drop=True, inplace=True)

#### 2.5 Encode track_id and user_id

# count unique track_id's in both dataframes
print('Number of unique track_id\'s in user_df: ', user_df['track_id'].nunique())
print('Number of unique track_id\'s in music_df: ', features_df['track_id'].nunique())

# encode track_id as numeric in both dataframes
track_id_to_numeric = {track_id: i for i, track_id in enumerate(features_df['track_id'])}

# double-check uniques
user_df['track_id_enc'] = user_df['track_id'].map(track_id_to_numeric)
features_df['track_id_enc'] = features_df['track_id'].map(track_id_to_numeric)

# sanity check: count unique track_id's in both dataframes... again
print('Number of unique track_id\'s in user_df: ', user_df['track_id_enc'].nunique())
print('Number of unique track_id\'s in music_df: ', features_df['track_id_enc'].nunique())


# test if encoding worked
# for track_id_enc = 10705, check if track_id in other dataframe has the same values
user_df[user_df['track_id_enc'] == 10705]['track_id'].unique() == features_df[features_df['track_id_enc'] == 10705]['track_id'].unique()

# unique user_id's in user_df
user_df['user_id'].nunique()

# encode user_id in user_df dataframe
user_df['user_id'] = user_df['user_id'].astype('category')
user_df['user_id_enc'] = user_df['user_id'].cat.codes

# verify unique user_id's in user_df
user_df['user_id'].nunique()

# replace original ids with encoded ids
user_df['track_id'] = user_df['track_id_enc']
user_df['user_id'] = user_df['user_id_enc']
features_df['track_id'] = features_df['track_id_enc']

# drop encoded columns
user_df.drop(columns=['track_id_enc', 'user_id_enc'], inplace=True)
features_df.drop(columns=['track_id_enc'], inplace=True)

### Wrapping up

user_df.head()

features_df.head()

# dataframe of track_id, name and artist
titles = features_df[['track_id', 'name', 'artist']]

# test numbers
len(features_df) == user_df['track_id'].nunique() == len(titles)

import pickle
pickle.dump(features_df, open('features.pkl', 'wb'))
pickle.dump(user_df, open('playcounts.pkl', 'wb'))
pickle.dump(titles, open('titles.pkl', 'wb'))
