==> 03-collaborative.ipynb <==
# Model training

import pandas as pd
import numpy as np
import pickle
from surprise import SVD, Dataset, Reader
from surprise.accuracy import rmse
from surprise.model_selection import train_test_split, cross_validate
from surprise.model_selection import GridSearchCV
from surprise import KNNWithMeans


playcounts = pd.read_pickle('playcounts.pkl')

playcounts.head()

# 0.1 sample of data
sample = playcounts.sample(frac=0.1, random_state=42)

# comment this out if you want to use the full dataset. sample is for gridsearc
#playcounts = sample

len(playcounts)

#### create surprise sets

reader = Reader()
data = Dataset.load_from_df(playcounts[['user_id', 'track_id', 'playcount']], reader)

### SVD

# parameter grid for SVD
svd_param_grid = {
    'n_factors': [40, 50],  # Number of factors
    'n_epochs': [30, 40, 50],     # Number of epochs
    'lr_all': [0.005, 0.01, 0.15],      # Learning rate
    'reg_all': [0.07, 0.1]        # Regularization term
}

# hyperparameter tuning for SVD
gs_svd = GridSearchCV(SVD, svd_param_grid, measures=['rmse', 'mae'], cv=3, n_jobs=3)

gs_svd.fit(data)

# best RMSE score
print(gs_svd.best_score['rmse'])
# best parameters
svd_best_params = gs_svd.best_params['rmse']
print(svd_best_params)

### KNN gridsearch

knn_param_grid = {
    'k': [9, 15, 19, 25],
    'sim_options': {
        'name': ['cosine', 'pearson_baseline'],
        'min_support': [5, 10, 15],
        'user_based': [False]
    }
}

gs_knn = GridSearchCV(KNNWithMeans, knn_param_grid, measures=['rmse', 'mae'], cv=3, n_jobs=3)

gs_knn.fit(data)

# best RMSE score
print(gs_knn.best_score['rmse'])
# best parameters
knn_best_params = gs_knn.best_params['rmse']
print(knn_best_params)

### test knn

k = 9 #knn_best_params['k']
sim_options = {'name': 'pearson_baseline', 'min_support': 15, 'user_based': False} #knn_best_params['sim_options']

trainset, testset = train_test_split(data, test_size=0.2)

# train with best parameters
knn = KNNWithMeans(k=k, sim_options=sim_options)

knn.fit(trainset)

predictions = knn.test(testset)

accuracy_rmse = rmse(predictions)

### Train the whole set

trainset = data.build_full_trainset()

knn = KNNWithMeans(k=k, sim_options=sim_options)

knn.fit(trainset)

# create mappings from track_id to inner id and vice versa
inner_to_track = {}
track_to_inner = {}
for inner_id in trainset.all_items():
    track_id = trainset.to_raw_iid(inner_id)
    inner_to_track[inner_id] = track_id
    track_to_inner[track_id] = inner_id



### knn innerid-rawid testing

testcase = 3

def get_neighbors(outer, k):
    neighbors = []
    inner = track_to_inner[outer]
    for inner_id in knn.get_neighbors(inner, k):
        neighbors.append(inner_to_track[inner_id])
    return neighbors

get_neighbors(testcase, 10)

### Re-order knn similarity matrix to match track_id order

# create knn_similarities matrix
knn_similarities = knn.compute_similarities()

knn_sim_ordered = np.copy(knn_similarities)

# rearrange vectors
for i in range(len(knn_similarities)):
    knn_sim_ordered[inner_to_track[i]] = knn_similarities[i]

# transpode the matrix for further processing
knn_sim_ordered = knn_sim_ordered.transpose()

knn_sim_ordered_2 = np.copy(knn_sim_ordered)

# rearrange vectors of transposed matrix
for i in range(len(knn_similarities)):
    knn_sim_ordered_2[inner_to_track[i]] = knn_sim_ordered[i]

(knn_sim_ordered_2.transpose() == knn_sim_ordered_2).all()

# house cleaning
del knn_similarities
del knn_sim_ordered

# convert float64 to float16 to save space
knn_sim_ordered_2 = np.array(knn_sim_ordered_2, dtype=np.float16)

# store model in pickle
pickle.dump(knn_sim_ordered_2, open('knn_matrix.pkl', 'wb'))
