==> 05-hybrid.ipynb <==
# Hybrid Recommender System
### combine content-based and collaborative filtering

import pandas as pd
import numpy as np
import pickle
from surprise import KNNWithMeans

# load cosine_matrix.pkl
with open('cosine_matrix.pkl', 'rb') as f:
    cosine_matrix = pickle.load(f)

# load knn.pkl
with open('knn_matrix.pkl', 'rb') as f:
    knn_matrix = pickle.load(f)
    

# load titles.pkl
with open('titles.pkl', 'rb') as f:
    titles = pickle.load(f)

#### weighted matrix

# change weights if desired
knn_sim_weight = 0.5
cos_sim_weight = 0.5

weighted_sim = (knn_sim_weight * knn_matrix) + (cos_sim_weight * cosine_matrix)

### Retrieving top recommendations

titles.head()

def get_track_id(artist, song):
    '''Get index from titles where artist and song match'''
    
    try:
        track_id = titles[(titles['artist'] == artist) & (titles['name'] == song)].index[0]
        return track_id
    except:
        print('Song not found in database')
        return None

# test
get_track_id('Oasis', 'Wonderwall')

def get_song(index):
    '''Get song artist and name at index'''
    try:
        row = titles.iloc[index]
        artist_song = row['artist'] + ' - ' + row['name']
        return artist_song
    except:
        print('Index not found in database')
        return None

# test
get_song(0)

def recommend_content(artist, title, sim_matrix):
    '''Get similar songs using the similarity matrix'''
    # get index for our song
    idx = get_track_id(artist, title)
    
    # get similarity scores of all songs w.r.t to our song
    sim_scores = list(enumerate(sim_matrix[idx]))
    
    # sort scores based on similarity
    sorted_sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)
    
    # limit to 20 songs
    sorted_sim_scores = sorted_sim_scores[1:20]

    # get song artist and names
    content_similar_scores = []
    for i in sorted_sim_scores:
        content_similar_scores.append(get_song(i[0]))
    
    return content_similar_scores

# select 10 random rows from titles and display artist and name
titles.sample(10)[['artist', 'name']]

# recommend_content test
recommend_content('Flying Lotus', 'SexSlaveShip', weighted_sim)

# store weighted_sim in pickle
with open('weighted_sim.pkl', 'wb') as f:
    pickle.dump(weighted_sim, f)
